Big diff this time. The biggest thing that's been done is the removal of as much static state as possible. This means removing static fields, such as VM::vm, GC::gc, Module::loadedModules, and others. They are now instead placed in the VM class, which is instantiated in VM_Start, and kept on the stack for the duration of the program.
The primary purpose of this change is to increase Ovum's isolation, so that (1) multiple Ovums can be started per process, (2) Ovum can be safely embedded in other programs. There is still a very long way to go before that goal is reached, but this change is an important step towards that.

As a result of these modifications, it is now also necessary for API functions to pass around some kind of object that can be asked for the current VM. In general, this means a lot of API functions have been modified to take a ThreadHandle where they previously did not. Hence, lots of little code changes in aves as well.

Coming up (eventually): a safe, easy-to-use way of declaring static fields in native modules. So that you don't have to use C++ static fields.

[VM]
* Removed various static fields, and put them in ovum::VM (as instance fields),
* Modified a lot of code to pass around an instance of ovum::VM, or a type that contains a reference to the VM instance.
* Changed plenty of classes so that they now access the owning VM instance (and/or CG and/or module pool and/or reference signature pool, etc.) through their own pointer, rather than the VM::vm field, which is now gone.
* All the module reading methods (ReadModuleRefs, ReadTypeRefs, ReadSingleType, etc.) are now instance methods on Module, rather than being static. 'this' contains the module that receives the loaded members.
* Extracted Module::Pool into an isolated class, ModulePool.
* Added a basic thread-local storage (TLS) implementation, in TlsEntry<T>. This will be used for storing various managed runtime data, which will eventually allow developers of native modules to e.g. obtain managed handles inside callbacks to unmanaged methods.
* Fixed a bug in Thread::GetStackTrace, which caused a segfault if any argument on the call stack was a reference to a null value.

[aves]
* Changed various parts of the C++ code to fit the signatures of the modified API functions. Generally, this just means passing an extra ThreadHandle as the first argument.
* Added documentation to members of aves.Method.
* Added operators >> and << to aves.Method, which allow single-arg Methods to be chained. See the documentation comments for details.
* Minor documentation fix for aves.String methods.

[aves.tests]
* Added extremely basic tests for aves.Char; more to come later.