[ovum-vm] Get rid of static LitStrings and String*s * Move static LitStrings and String*s to StaticStrings * Update all affected code to use StaticStrings/VM->GetStrings() * Add a script for generating the StaticStrings data from a JSON file * Move standard type info into its own classes, StandardTypeInfo and StandardTypeCollection * Update code to use StandardTypeInfo/StandardTypeCollection
This is quite a big change, but it is definitely a necesary one. Those parts of the code that use LitString<> to initialise static string values, such as names of standard types and special members like ".new", have been entirely reworked. There is now a single, unified interface for all such strings: ovum::StaticString, which ovum::VM keeps an instance of.

One of the major reasons for doing this is to do with string hashing. For security reasons, Ovum should one day be given support for randomised string hashes, which in practice means string hashes modified by a random seed initialised when the VM starts. Static strings would by necessity share the same hash code across all instances of Ovum in a single process.

A similar change is planned for aves, which will require a slightly different solution.

Note: the data in StaticStrings is still initialised from static values, but a private copy is made by the VM instance during startup.