[aves] Prohibit negative indexes in all List members
It is no longer possible to access list items by negative index, e.g. `list[-1]` to get the last item. In general, the feature is not especially useful, and it's a bit of a mess to maintain and implement consistently everywhere. It is more useful to think of the range of valid indexes as [0, length), in which case a negative index is clearly an error. Indeed it seems that, in code generally, negative indexes indicate an error in a calculation.

Hence, negative indexes are gone.