Change concatenation behaviour [see below fold]
The concatenation operator, '::', is changed so that it only concatenates strings. The ability to combine lists and hash tables with this operator is removed. However, since that functionality is sometimes useful, List and Hash both get a concat() method, which does precisely what '::' used to do.

The compiler will be altered to treat the result of a concatenation-expression as having the known type aves.String.