[aves] Move higher-order functions to Iterable
The type Iterator is not really meant to be used by developers under most circumstances. It's a support type; it's used in for-in statements, and maybe occasionally implemented for a new collection type. The purpose of Iterator is to give you the contents of a sequence. All manipulations to the sequence is done on Iterable (or Collection) instances, and the developer should be encouraged to work with that type instead.

Moreover, duplicating functionality across Iterator and Iterable leads to confusion: am I supposed to override methods in Iterable or Iterator (if my new collection type has a more optimized implementation)? Plus, documentation must be synchronized across two files, which is guaranteed to result in inconsistencies.

To keep Iterator streamlined, all of the higher-order functionality is removed from it. However, toList() and toArray() remain, as they are very useful utility methods.

Interestingly, no published code (outside Iterable, internal things and tests) is affected by this change. Everything uses the methods in Iterable. This is therefore considered a safe move.

Hurrah for cleaning up.