[aves] Move sequence sorting functionality to Iterable
In this change, List.sort() is removed, and Iterable gains the methods sort() and sortBy(). Note that this removes the ability to pass in a custom comparer function; instead, you pass in one or more key selector functions. This is generally a much more preferable design, as sort keys can be cached in various nice ways, and it is far more intuitive to use, and far less error-prone.

This change breaks existing code that relies on List.sort(comparer).