Change opcodes for operators
Now that 'hashop' and 'dollar' (the two 'unreserved operators') are gone, there is no need to reserve space for them in the middle of the operators block. To avoid having to rewrite too much code, however, we move the two reserved slots down a bit instead of adjusting later instructions.