Implement support for fault blocks
A fault block can be described in two ways: it's either like a finally
block that only gets executed when an error is thrown; or it's like a
catch that doesn't actually handle the error.

Basically: when a thrown error attempts to propagate past the fault block,
the block is run much like a catch; but it does not receive the error on
the stack and has no effect on error propagation, much like a finally.
Another important difference between a fault block and catch+rethrow is
that a fault block retains the original stack trace. While a catch block
*handles* an error, a fault block merely *reacts* to it.

There is no Osprey syntax that corresponds directly to fault blocks, and
the compiler does not currently make use of them. However, they will be
used in the future to implement state machines.