Normalize integral type use, especially size_t
Ovum has been highly inconsistent with its use of integral types,
particularly `size_t`. For things that hold a count of things, Ovum has
used a wildly varying mixture of `int32_t`, `uint32_t`, `int`, `unsigned
int`, `uint16_t` (mostly in the past), and only occasionally `size_t`.

This commit attempts to remedy the problem.

First and foremost: Fields and variables that contain a count of things
are now of type `size_t`. Likewise all methods that return values that
represent counts. This includes internal things as well as the external
API, Ovum as well as aves. Needless to say, this is a massive change. Some
algorithms, such as looking up entries in hash tables, had to be modified.

Secondly: In some cases it is actually necessary to return negative
numbers, for example to indicate error conditions. In these cases we use
`ssize_t` instead. This type is actually POSIX-specific, but the Windows
header typedefs it to `SSIZE_T`, which serves the same purpose in
Win32-land. The `ssize_t` type is used sparingly, and never in any public
API functions; only internally in Ovum and aves.

Thirdly: While Ovum (now) uses `size_t`, which is unsigned, Osprey-facing
APIs in aves prefer to return `aves.Int`-values, which are signed. On
32-bit every `size_t` fits inside `int64_t`; on 64-bit, especially large
values of `size_t` do not fit inside `int64_t`. For this reason, the Ovum
headers introduce the constant `OVUM_ISIZE_MAX`, which is used as the
actual maximum capacity of lists, hashes, sets, arrays, buffers and other
collection types that may arise. On 32-bit, it's `SIZE_MAX`; on 64-bit,
it's `INT64_MAX`. The latter (9223 petabytes) should be enough for any
reasonable uses of Ovum, and quite a lot of unreasonable uses too.

*Unfortunately,* C++ is entirely too fond of converting implicitly between
all sorts of numeric types, whether the conversion is lossy or not, in
various arithmetic operations, sometimes including comparisons. As a
result, it is quite likely (in fact, virtually guaranteed) that I have
missed several incorrect uses of integral types. These will be fixed as
they are discovered.

Lastly, a variety of other fixes - chiefly formatting and memory
management techniques (i.e. changing to ovum::Box in places) - have
managed to sneak into this commit. Oh well.