BREAKING: Only null and false are falsy now!
Previously, the following values were considered falsy:

1. Null
2. Any value of a primitive type whose bytes are all 0. This includes
   false, 0, 0u, 0.0 (but see below), '\0', and any enum value with the
   underlying integral value 0.

While these criteria had the benefit of being very fast to check, they
were not without problems:

1. It is somewhat arbitrary. Why should enum values with an underlying
   value of 0 be false? Normally we even try to hide the underlying value;
   this behaviour is potentially surprising.
2. It was kind of an implementation cheat. More to the point, the Osprey
   spec was designed to make Ovum in particular easy to implement. This is
   never okay.
3. Floating-point is problematic. To implement the Osprey spec correctly,
   we would have to make -0.0 falsy too. And what about NaN? Many
   environments (sensibly) treat NaN as falsy, but not Osprey
4. More on arbitrariness: Why _aren't_ some values falsy? Like `""`, `[]`,
   `{}`, and other empty collections. In writing this commit I initially
   forgot about '\0', too. This indicates the rules may be too complex.

In general, when checking something that isn't null or false, it is better
to use more explicit conditions: `length > 0`, `foo != ""`, `bar ==
Bar.none`, and so on.

So, from now on, only false and null are falsy. Null remains falsy for two
reasons: (1) it represents the absence of a meaningful value, an empty
box, an unused slot; it's not truthy in the same way that "off" is not a
TV channel; (2) null-like values are traditionally treated as falsy in
other languages, and being consistent with that in Osprey probably
generally reduces surprise. (One can have a variety of mostly valid
opinions on the uselessness of having null in the first place, but that's
a discussion for another day...)

Remarkably, this change required almost no modifications to existing
Osprey code. Apparently it already consistently uses explicit condition
forms.