Make RootSetWalker and ObjecGraphWalker template types
The only real benefit of turning these into template types is that it
allows us to get rid of the virtual calls on the Visitor types. Since
the GC is full of hot paths, avoiding virtual calls during a GC cycle
seems like a good idea.

Due to the way C++ works, this does of course (unfortunately) involve
putting all the actual walking code inside the header files.